projeto_bula/requirements.txt
streamlit
langchain
langchain_openai
langchain-core
langchain-community
python-dotenv
pypdf
chromadb